add util tests to util file
